[refactor] splitBySpace Shell -> util 위치 변경, header/cpp 분리
