+++ TestShell/TestShell.vcxproj.filters
@@ -57,8 +57,8 @@
     <ClCompile Include="util.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="ICmd.h">
-      <Filter>헤더 파일</Filter>
+    <ClCompile Include="Shell.cpp">
+      <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
