+++ TestShell/util.cpp
@@ -1,5 +1,7 @@
 #include "util.h"
 #include <stdexcept>
+#include <vector>
+#include <sstream>
 
 bool startsWith(const std::string& str, const std::string& prefix) {
 	return str.size() >= prefix.size() &&
@@ -14,4 +16,20 @@ bool isOutOf4ByteRange(const std::string& hexStr) {
     catch (const std::exception&) {
         return true;
     }
+}
+
+std::vector<std::string> splitBySpace(const std::string& input) {
+	std::istringstream iss(input);
+	std::vector<std::string> result;
+	std::string word;
+
+	if (input.empty()) {
+		return result;
+	}
+
+	while (iss >> word) {
+		result.push_back(word);
+	}
+
+	return result;
 }
